Rename URI handler files
